Detail Screen Implementation

+++ core/domain/src/main/java/com/example/movieapp/domain/usecase/movie/ToggleFavouriteUseCase.kt
@@ -1,5 +1,6 @@
 package com.example.movieapp.domain.usecase.movie
 
+import com.example.movieapp.domain.model.movie.MovieDetail
 import com.example.movieapp.domain.model.movie.PopularMovie
 import com.example.movieapp.domain.repository.movie.FavouriteMovieRepository
 import kotlinx.coroutines.flow.first
@@ -7,12 +8,25 @@ import kotlinx.coroutines.flow.first
 class ToggleFavouriteUseCase(
     private val repository: FavouriteMovieRepository
 ) {
-    suspend operator fun invoke(movie: PopularMovie){
+    suspend operator fun invoke(movie: PopularMovie) {
+        toggle(movie)
+    }
+
+    suspend operator fun invoke(movie: MovieDetail) {
+        toggle(
+            PopularMovie(
+                id = movie.id,
+                title = movie.title,
+                posterUrl = movie.posterUrl,
+                year = movie.releaseYear,
+                isFavorite = movie.isFavorite,
+                category = ""
+            )
+        )
+    }
+
+    private suspend fun toggle(movie: PopularMovie) {
         val isCurrentlyFav = repository.isMovieFavourite(movie.id).first()
-        if(isCurrentlyFav){
-            repository.deleteFavourite(movie)
-        }else{
-            repository.insertFavourite(movie)
-        }
+        if (isCurrentlyFav) repository.deleteFavourite(movie) else repository.insertFavourite(movie)
     }
 }
